first commit in matmult

+++ matmult/src/utils/utils.cpp
@@ -1,6 +1,6 @@
 #include "./utils.h"
 
-std::pair <size_t, size_t> getOptimalGeometry(int n) {
+std::pair <int, int> getOptimalGeometry(int n) {
     std::pair<size_t, size_t> result;
     for (int i = 1; i * i < n; ++i) {
         if (n % i == 0) {
